Add tests for Console::writeLine(), fix typo in path.
