+++ tests/ZendTest/Console/TestAssets/ConsoleAdapter.php
@@ -20,6 +20,10 @@ class ConsoleAdapter extends AbstractAdapter
 
     public $autoRewind = true;
 
+    public $testWidth = 80;
+
+    public $testIsUtf8 = true;
+
     /**
      * Read a single line from the console input
      *
@@ -50,5 +54,36 @@ class ConsoleAdapter extends AbstractAdapter
             $char = fread($this->stream, 1);
         } while ("" === $char || ($mask !== null && false === strstr($mask, $char)));
         return $char;
+    }
+
+    /**
+     * Force reported width for testing purposes.
+     *
+     * @param int $width
+     * @return int
+     */
+    public function setTestWidth($width)
+    {
+        $this->testWidth = $width;
+    }
+
+    /**
+     * Force reported utf8 capability.
+     *
+     * @param bool $isUtf8
+     */
+    public function setTestUtf8($isUtf8)
+    {
+        $this->testIsUtf8 = $isUtf8;
+    }
+
+    public function isUtf8()
+    {
+        return $this->testIsUtf8;
+    }
+
+    public function getWidth()
+    {
+        return $this->testWidth;
     }
 }
